requirements.txt
Django==4.2
djangorestframework==3.14.0
markdown==3.4.3
django-filter==23.1
dj-rest-auth==3.0.0
drf-spectacular==0.26.1
pillow==9.5.0
django-allauth==0.52.0
gunicorn==20.1.0